refactor: move paging and patch dtos to base package
