+++ trends/src/main/kotlin/net/thechance/trends/api/dto/base/PagingResponse.kt
@@ -1,4 +1,4 @@
-package net.thechance.trends.api.dto
+package net.thechance.trends.api.dto.base
 
 data class PagingResponse<T>(
     val pageNumber: Int,
