+++ trends/src/main/kotlin/net/thechance/trends/api/dto/base/PatchMetadata.kt
@@ -1,4 +1,4 @@
-package net.thechance.trends.api.dto
+package net.thechance.trends.api.dto.base
 
 data class PatchMetadata(
     val addedCount: Int,
