+++ trends/src/main/kotlin/net/thechance/trends/api/dto/category/PatchUserCategoriesResponse.kt
@@ -1,6 +1,6 @@
 package net.thechance.trends.api.dto.category
 
-import net.thechance.trends.api.dto.PatchMetadata
+import net.thechance.trends.api.dto.base.PatchMetadata
 
 data class PatchUserCategoriesResponse(
     val patchMetadata: PatchMetadata,
